moved app.js to root of client folder
